pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
